feat: 텍스트 수정 컴포넌트 분리
